Align has template functions
