+++ src/data/template.ts
@@ -1,1 +1,18 @@
-export const isDynamicTemplate = (value: string) => value.includes("{{");
+const isTemplateRegex = new RegExp("{%|{{|{#");
+
+export const isTemplate = (value: string): boolean =>
+  isTemplateRegex.test(value);
+
+export const hasTemplate = (value: unknown): boolean => {
+  if (!value) {
+    return false;
+  }
+  if (typeof value === "string") {
+    return isTemplate(value);
+  }
+  if (typeof value === "object") {
+    const values = Array.isArray(value) ? value : Object.values(value!);
+    return values.some((val) => hasTemplate(val));
+  }
+  return false;
+};
