+++ src/panels/config/automation/action/types/ha-automation-action-delay.ts
@@ -7,11 +7,11 @@ import {
   PropertyValues,
 } from "lit-element";
 import { fireEvent } from "../../../../../common/dom/fire_event";
-import { hasTemplate } from "../../../../../common/string/has-template";
 import "../../../../../components/entity/ha-entity-picker";
 import { HaFormTimeData } from "../../../../../components/ha-form/ha-form";
 import "../../../../../components/ha-service-picker";
 import { DelayAction } from "../../../../../data/script";
+import { hasTemplate } from "../../../../../data/template";
 import { HomeAssistant } from "../../../../../types";
 import { ActionElement } from "../ha-automation-action-row";
